[C8] Split serialize_3d.go per entity family (#2243)
model/sketch/serialize_3d.go was 715 LOC, 1.4x the 500-line ceiling. Split
the 3D-sketch YAML serialization per entity family; each file is now under the
ceiling:

- serialize_3d.go (365): the DTOs, the top-level Marshal/Apply dispatch, the
  point serialize/restore, and the shared codec helpers (axisTriple,
  flatten/unflatten, plane-name/normal, entity/point lookups).
- serialize_3d_entity.go (219): the curve/entity family — the serialize/
  restore of lines, circles, conics, splines and helical curves (helix
  definition + spline-handle encoding + entity lookups).
- serialize_3d_constraint.go (154): the constraint family — the dimensional
  and geometric constraint serialize/restore (dimension, bend, helical).

Behaviour-preserving: split programmatically by receiver-qualified
declaration, verified every non-comment code line is byte-identical (only
per-file import redistribution differs); imports resolved per file with
goimports. serialize_*.go is already funlen/gocyclo-excluded by path, so no
config change is needed. Sketch package builds, archguard passes, lint 0.
Net delta of recognizers / tolerance constants / fallback sites / type
assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2243.

(cherry picked from commit 80482699a1c1285ea5819d8de4ced59334fd5de3)
